requirements.txt
langchain
langchain-openai
python-decouple
openai
langserve[all]
langchain-community
upstash-vector
jupyter
langchain-google-genai
google-generativeai
python-dotenv
tiktoken
fastapi
uvicorn
